+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Added `Wizaplace\SDK\Basket\Payment::getExternalReference`
 - Added `Wizaplace\SDK\Order\Payment::getExternalReference`
 - Added `Wizaplace\SDK\Catalog::getProductsByMvpId`
-- Added parameter `bool $allowMvp = true` to `Wizaplace\SDK\Catalog::getProductsBy` methods
+- Added parameter `bool $allowMvp = true` to `Wizaplace\SDK\Catalog::getProductsBy*` methods
 
 ## 1.67.2