+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next release
 
+## Bugfixes
+
+- Fixed `Vendor/Promotion/BasketPromotion::getRule` can return null
+
 ### New features
 
 - Added `Wizaplace\SDK\Catalog\Product::isMVP` to know if a product was a Multi-Vendor product or not